server: move status logic to statusApi
